bilbo contribution!! Mummi contribution!! veriverse contribution!! LCTV contribution!! triforcelink and gigabytegiant streamer contribution!! Catherina Cortega, Benjamin Kuhl, Pawel Kulesza, Michael J. Garbade, marianyc, kissloryshy, kord, kranalex and matroskinb contribution!! Discord contribution!! Catherina Cortega, r00tkiller and GigabyteGiant contribution!! aureliokessler, dopedopedope, dzetko, guitago, rex2, wilson2000 contribution!! djoga and guitago presence contribution!! Twitch contribution!! Beam contribution!!
Thank you Carlos!!

Thank you Mummi!!

Carol contribution!!

at bilbo at veriverse!!
i7 6700 16GB RAM SSD OCZ 120GB SSD 240GB HD 2TB Windows 10 Pro

+++ axis/pixman/pixman/pixman-gradient-walker.c
@@ -24,9 +24,9 @@
  */
 
 #ifdef HAVE_CONFIG_H
-//#include<config.h>
+#include <config.h>
 #endif
-//#include"pixman-private.h"
+#include "pixman-private.h"
 
 void
 _pixman_gradient_walker_init (pixman_gradient_walker_t *walker,
@@ -54,7 +54,7 @@ static void
 gradient_walker_reset (pixman_gradient_walker_t *walker,
 		       pixman_fixed_48_16_t      pos)
 {
-    int32_t x, left_x, right_x;
+    int64_t x, left_x, right_x;
     pixman_color_t *left_c, *right_c;
     int n, count = walker->num_stops;
     pixman_gradient_stop_t *stops = walker->stops;
